add deployment config cmd
